requirements.txt
pandas
numpy
scikit-learn
joblib
tensorflow
python-binance
ccxt
schedule
pytz
